fix: move growthbook set attributes directly in admin nav bar
